Set the WebAuthN rpId to the COOKIE_DOMAIN if set
By default the WebAuthN Passkey is scoped to the origin domain, for example, example.org

If a Multisite install however shares it's cookies with subdomains, attempting to login via subdomain.example.org will fail to match as the origin doesn't match the keys ID.

On single-site, COOKIE_DOMAIN is normally false. On Multisite, it's usually set to either a) The domain, or b) on subdomains, the parent domain. The ltrim() is used for sites that support older browsers which needed the leading `.`.

This allows for the WebAuthN Passkey to be setup via a subdomain or the parent domain, and used on all of it's subdomains or parent domain.